Loaded 400K samples to ADC!
I could not see the results, but the program loaded without issues. The problem seems to be the Callback counts. It was set to 2000. I tried to decrease it to 200 or less and the results were all 0s. Because of this number, the code actually executes 2K times. So if you work with 400K samples you are accessing the data 400K * 2K * 2 (channels)